PCM-1499 fixed existing tests (still need tests for media class)
