+++ test/unit/sessions/session-manager.test.ts
@@ -2,8 +2,7 @@ import { GenesysCloudWebrtcSdk } from '../../../src/client';
 import { SessionManager } from '../../../src/sessions/session-manager';
 import { SimpleMockSdk, createPendingSession, MockSession, createSessionInfo, MockStream, MockTrack } from '../../test-utils';
 import { SessionTypes } from '../../../src/types/enums';
-import { IUpdateOutgoingMedia, IExtendedMediaSession, IEnumeratedDevices } from '../../../src/types/interfaces';
-import * as mediaUtils from '../../../src/media-utils';
+import { IUpdateOutgoingMedia, IExtendedMediaSession, SdkMediaState } from '../../../src/types/interfaces';
 import BaseSessionHandler from '../../../src/sessions/base-session-handler';
 
 let mockSdk: GenesysCloudWebrtcSdk;
@@ -529,13 +528,13 @@ describe('updateOutputDeviceForAllSessions()', () => {
     const screenShareSession = new MockSession(SessionTypes.acdScreenShare);
     const videoSession = new MockSession(SessionTypes.collaborateVideo);
 
-    jest.spyOn(mediaUtils, 'getValidDeviceId').mockResolvedValue(undefined);
+    jest.spyOn(mockSdk.media, 'getValidDeviceId').mockReturnValue(undefined);
     jest.spyOn(sessionManager, 'getAllActiveSessions').mockReturnValue([screenShareSession, videoSession] as any);
     jest.spyOn(sessionManager, 'getSessionHandler');
 
     await sessionManager.updateOutputDeviceForAllSessions(outputDeviceId);
 
-    expect(mediaUtils.getValidDeviceId).toHaveBeenCalledWith(mockSdk, 'audiooutput', outputDeviceId, videoSession);
+    expect(mockSdk.media.getValidDeviceId).toHaveBeenCalledWith('audiooutput', outputDeviceId, videoSession);
     expect(sessionManager.getSessionHandler).not.toHaveBeenCalled();
     expect(mockSdk.logger.warn).toHaveBeenCalledWith(
       expect.stringContaining('Output deviceId not found. Not updating output media'),
@@ -555,13 +554,13 @@ describe('updateOutputDeviceForAllSessions()', () => {
     ];
     const mockSessionHandler = { updateOutputDevice: jest.fn() };
 
-    jest.spyOn(mediaUtils, 'getValidDeviceId').mockResolvedValue(outputDeviceId);
+    jest.spyOn(mockSdk.media, 'getValidDeviceId').mockReturnValue(outputDeviceId);
     jest.spyOn(sessionManager, 'getAllActiveSessions').mockReturnValue(sessions as any);
     jest.spyOn(sessionManager, 'getSessionHandler').mockReturnValue(mockSessionHandler as any);
 
     await sessionManager.updateOutputDeviceForAllSessions(outputDeviceId);
 
-    expect(mediaUtils.getValidDeviceId).toHaveBeenCalledWith(mockSdk, 'audiooutput', outputDeviceId, sessions[0], sessions[1]);
+    expect(mockSdk.media.getValidDeviceId).toHaveBeenCalledWith('audiooutput', outputDeviceId, sessions[0], sessions[1]);
 
     sessions
       .filter(session => session.sessionType !== SessionTypes.acdScreenShare)
@@ -574,13 +573,13 @@ describe('updateOutputDeviceForAllSessions()', () => {
 
 describe('validateOutgoingMediaTracks()', () => {
   let sessions: IExtendedMediaSession[];
-  let devices: IEnumeratedDevices;
   let mockSessionHandler: BaseSessionHandler;
-
+  let testDevices: MediaDeviceInfo[];
+  let mediaState: SdkMediaState;
   let mockGetAllSessions: jest.SpyInstance<IExtendedMediaSession[]>;
-  let mockGetEnumeratedDevices: jest.SpyInstance<Promise<IEnumeratedDevices>>;
   let mockGetSession: jest.SpyInstance<IExtendedMediaSession>;
   let mockGetSessionHandler: jest.SpyInstance<BaseSessionHandler>;
+  let setMediaStateDevices: typeof mockSdk.media['setDevices'];
 
   beforeEach(() => {
     mockSessionHandler = {
@@ -594,20 +593,20 @@ describe('validateOutgoingMediaTracks()', () => {
     sessions = [
       new MockSession() as any
     ];
-    devices = {
-      audioDevices: [
-        { deviceId: 'device#1', groupId: 'group#1', kind: 'audioinput', label: 'Device #1' } as any,
-        { deviceId: 'device#2', groupId: 'group#2', kind: 'audioinput', label: 'Device #2' } as any
-      ],
-      videoDevices: [
-        { deviceId: 'device#3', groupId: 'group#3', kind: 'videoinput', label: 'Device #3' } as any,
-        { deviceId: 'device#4', groupId: 'group#4', kind: 'videoinput', label: 'Device #4' } as any
-      ],
-      outputDevices: []
-    };
+    testDevices = [
+      /* audioDevices */
+      { deviceId: 'device#1', groupId: 'group#1', kind: 'audioinput', label: 'Device #1' } as any,
+      { deviceId: 'device#2', groupId: 'group#2', kind: 'audioinput', label: 'Device #2' } as any,
+      /* videoDevices */
+      { deviceId: 'device#3', groupId: 'group#3', kind: 'videoinput', label: 'Device #3' } as any,
+      { deviceId: 'device#4', groupId: 'group#4', kind: 'videoinput', label: 'Device #4' } as any
+    ];
+
+    setMediaStateDevices = mockSdk.media['setDevices'].bind(mockSdk.media);
 
     mockGetAllSessions = jest.spyOn(sessionManager, 'getAllActiveSessions').mockReturnValue(sessions);
-    mockGetEnumeratedDevices = jest.spyOn(mediaUtils, 'getEnumeratedDevices').mockResolvedValue(devices);
+    setMediaStateDevices(testDevices);
+    mediaState = mockSdk.media.getState();
 
     mockGetSession = jest.spyOn(sessionManager, 'getSession').mockImplementation(({ id }) => sessions.find(s => s.id === id));
     mockGetSessionHandler = jest.spyOn(sessionManager, 'getSessionHandler').mockReturnValue(mockSessionHandler as any);
@@ -615,7 +614,7 @@ describe('validateOutgoingMediaTracks()', () => {
 
   it('should ignore the screen share stream on the session', async () => {
     const screenShareStream = new MockStream({ video: true });
-    const mockTrack = new MockTrack('video', devices.videoDevices[0].label);
+    const mockTrack = new MockTrack('video', mediaState.videoDevices[0].label);
     const session = sessions[0];
 
     session._screenShareStream = screenShareStream as any;
@@ -630,11 +629,10 @@ describe('validateOutgoingMediaTracks()', () => {
 
     /* other generic mocks called */
     expect(mockGetAllSessions).toBeCalled();
-    expect(mockGetEnumeratedDevices).toBeCalled();
   });
 
   it('should not update the session if the media did not get lost', async () => {
-    const mockTrack = new MockTrack('video', devices.videoDevices[0].label);
+    const mockTrack = new MockTrack('video', mediaState.videoDevices[0].label);
     const session = sessions[0];
     session.pc['_addSender'](mockTrack); /* this is a mock PC */
 
@@ -647,15 +645,16 @@ describe('validateOutgoingMediaTracks()', () => {
   });
 
   it('should update video media on all sessions if the device was lost', async () => {
-    const mockVideoTrack = new MockTrack('video', devices.videoDevices[0].label);
-    const mockAudioTrack = new MockTrack('audio', devices.audioDevices[0].label);
+    const mockVideoTrack = new MockTrack('video', mediaState.videoDevices[0].label);
+    const mockAudioTrack = new MockTrack('audio', mediaState.audioDevices[0].label);
 
     const session = sessions[0];
     session.pc['_addSender'](mockVideoTrack); /* this is a mock PC */
     session.pc['_addSender'](mockAudioTrack);
 
-    /* device is removed */
-    devices.videoDevices = devices.videoDevices.slice(1);
+    /* video device is lost */
+    const newDevices = testDevices.filter(d => d.label !== mediaState.videoDevices[0].label);
+    setMediaStateDevices(newDevices);
 
     mockSdk.logger.warn['mockReset']();
     await sessionManager.validateOutgoingMediaTracks();
@@ -679,15 +678,16 @@ describe('validateOutgoingMediaTracks()', () => {
   });
 
   it('should update audio media on all sessions if the device was lost', async () => {
-    const mockVideoTrack = new MockTrack('video', devices.videoDevices[0].label);
-    const mockAudioTrack = new MockTrack('audio', devices.audioDevices[0].label);
+    const mockVideoTrack = new MockTrack('video', mediaState.videoDevices[0].label);
+    const mockAudioTrack = new MockTrack('audio', mediaState.audioDevices[0].label);
 
     const session = sessions[0];
     session.pc['_addSender'](mockVideoTrack); /* this is a mock PC */
     session.pc['_addSender'](mockAudioTrack);
 
-    /* device is removed */
-    devices.audioDevices = devices.audioDevices.slice(1);
+    /* audio device is lost */
+    const newDevices = testDevices.filter(d => d.label !== mediaState.audioDevices[0].label);
+    setMediaStateDevices(newDevices);
 
     mockSdk.logger.warn['mockReset']();
     await sessionManager.validateOutgoingMediaTracks();
@@ -707,8 +707,8 @@ describe('validateOutgoingMediaTracks()', () => {
   });
 
   it('should mute the media type if there is no device to switch to', async () => {
-    const mockVideoTrack = new MockTrack('video', devices.videoDevices[0].label);
-    const mockAudioTrack = new MockTrack('audio', devices.videoDevices[0].label);
+    const mockVideoTrack = new MockTrack('video', mediaState.videoDevices[0].label);
+    const mockAudioTrack = new MockTrack('audio', mediaState.videoDevices[0].label);
 
     const session = sessions[0];
     session.pc['_addSender'](mockVideoTrack); /* this is a mock PC */
@@ -716,8 +716,7 @@ describe('validateOutgoingMediaTracks()', () => {
     session._outboundStream = { removeTrack: jest.fn() } as any;
 
     /* all devices are removed (none available to switch to) */
-    devices.videoDevices = [];
-    devices.audioDevices = [];
+    setMediaStateDevices([]);
 
     await sessionManager.validateOutgoingMediaTracks();
 
@@ -757,11 +756,11 @@ describe('validateOutgoingMediaTracks()', () => {
     const session = sessions[0];
     session._outputAudioElement = mockOutputElement as any;
 
-    devices.outputDevices = [
-      { deviceId: 'some-device-id' } as MediaDeviceInfo
-    ];
+    setMediaStateDevices([
+      { deviceId: 'some-device-id', kind: 'audiooutput' } as MediaDeviceInfo
+    ]);
 
-    jest.spyOn(mediaUtils, 'hasOutputDeviceSupport').mockReturnValue(true);
+    mockSdk.media['setStateAndEmit']({ hasOutputDeviceSupport: true }, 'state');
     jest.spyOn(sessionManager, 'updateOutputDeviceForAllSessions').mockResolvedValue(null);
 
     await sessionManager.validateOutgoingMediaTracks();
@@ -774,7 +773,7 @@ describe('validateOutgoingMediaTracks()', () => {
     expect(sessionManager.updateOutputDeviceForAllSessions).not.toHaveBeenCalled();
 
     /* device does not exist, attempt to switch */
-    devices.outputDevices = [];
+    setMediaStateDevices([]);
     await sessionManager.validateOutgoingMediaTracks();
 
     expect(mockSdk.logger.info).toHaveBeenCalledWith(
